requirements.txt
certifi==2019.11.28
chardet==3.0.4
idna==2.8
pg8000==1.13.2
Pillow==6.2.0
python-dotenv==0.10.4
pytz==2019.2
requests==2.22.0
scramp==1.1.0
six==1.14.0
slackbot==0.5.6
slacker==0.13.0
urllib3==1.25.7
websocket-client==0.44.0
